Move TestConfigConsolidation() and its helpers to a separate file
